requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
boto3==1.34.61
botocore==1.34.61
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
environs==11.0.0
idna==3.6
jmespath==1.0.1
marshmallow==3.21.1
oauthlib==3.2.2
packaging==24.0
paramiko==3.4.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
